requirements.txt
requests
dnspython
aiohttp
